pyproject.toml
[project]
name = "idwriter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "instructor>=1.8.3",
    "openpyxl>=3.1.5",
    "portkey-ai>=1.13.0",
    "pydantic>=2.11.5",
    "python-dotenv>=1.1.0",
    "pyyaml>=6.0.2",
]
